Move model-like classes to Models
Some of the stuff I had in Controllers seemed more appropriate to move into Models
